+++ src/main/java/eureca/capstone/project/batch/common/schedule/TransactionStatisticScheduler.java
@@ -1,0 +1,37 @@
+package eureca.capstone.project.batch.common.schedule;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.JobParametersBuilder;
+import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class TransactionStatisticScheduler {
+
+    private final JobLauncher jobLauncher;
+    private final Job transactionStatisticJob;
+
+    @Scheduled(cron = "0 0 * * * *", zone = "Asia/Seoul")
+    public void runStatisticBatch() {
+        try {
+            log.info("[runStatisticBatch] 거래량/시세 통계 배치 실행");
+            jobLauncher.run(transactionStatisticJob, new JobParametersBuilder()
+                    .addLong("time", System.currentTimeMillis())
+                    .addString("currentTime", LocalDateTime.now().toString())
+                    .toJobParameters());
+            log.info("[runStatisticBatch] 거래량/시세 통계 배치 실행 완료");
+        } catch (Exception e) {
+            log.error("[runStatisticBatch] 거래량/시세 통계 배치 실행 실패", e);
+        }
+    }
+}
+
