+++ src/main/java/eureca/capstone/project/batch/component/external/slack/config/WebClientConfig.java
@@ -1,0 +1,17 @@
+package eureca.capstone.project.batch.component.external.slack.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.builder()
+                .baseUrl("https://slack.com/api")
+                .defaultHeader("Content-Type", "application/json")
+                .build();
+    }
+}
