+++ src/main/java/eureca/capstone/project/batch/component/external/slack/dto/SlackMessageRequestDto.java
@@ -1,0 +1,19 @@
+package eureca.capstone.project.batch.component.external.slack.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+public class SlackMessageRequestDto {
+    String jobName;
+    String stepName;
+    Long readCount;
+    Long writeCount;
+    LocalDateTime startTime;
+    LocalDateTime endTime;
+    String exitSatus;
+    String errorMessage;
+}
