+++ src/main/java/eureca/capstone/project/batch/component/external/slack/service/SlackService.java
@@ -1,0 +1,9 @@
+package eureca.capstone.project.batch.component.external.slack.service;
+
+import eureca.capstone.project.batch.component.external.slack.dto.SlackMessageRequestDto;
+
+public interface SlackService {
+    void sendMessageToChannel(String channel, SlackMessageRequestDto message);
+
+    void sendMessage(SlackMessageRequestDto message);
+}
