+++ src/main/java/eureca/capstone/project/batch/component/listener/BatchFailureHelper.java
@@ -3,8 +3,8 @@ package eureca.capstone.project.batch.component.listener;
 import eureca.capstone.project.batch.auth.entity.UserAuthority;
 import eureca.capstone.project.batch.common.entity.BatchFailureLog;
 import eureca.capstone.project.batch.common.service.BatchFailureLogService;
-import eureca.capstone.project.batch.component.external.DiscordNotificationService;
-import eureca.capstone.project.batch.config.AuctionJobConfig;
+import eureca.capstone.project.batch.component.external.discord.DiscordNotificationService;
+import eureca.capstone.project.batch.job.AuctionJobConfig;
 import eureca.capstone.project.batch.transaction_feed.entity.TransactionFeed;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
