+++ src/main/java/eureca/capstone/project/batch/component/listener/ExecutionListener.java
@@ -1,8 +1,8 @@
 package eureca.capstone.project.batch.component.listener;
 
-import eureca.capstone.project.batch.component.external.DiscordNotificationService;
-import eureca.capstone.project.batch.slack.dto.SlackMessageRequestDto;
-import eureca.capstone.project.batch.slack.service.SlackService;
+import eureca.capstone.project.batch.component.external.discord.DiscordNotificationService;
+import eureca.capstone.project.batch.component.external.slack.dto.SlackMessageRequestDto;
+import eureca.capstone.project.batch.component.external.slack.service.SlackService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.*;
