--- src/main/java/eureca/capstone/project/batch/slack/config/WebClientConfig.java
@@ -1,17 +1,0 @@
-package eureca.capstone.project.batch.slack.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    public WebClient webClient() {
-        return WebClient.builder()
-                .baseUrl("https://slack.com/api")
-                .defaultHeader("Content-Type", "application/json")
-                .build();
-    }
-}
